# server/swagger.yaml
openapi: 3.0.0
info:
  title: Node.js + MySQL - Email Sign Up API
  description: Node.js + MySQL API for Email Sign Up with Verification, Authentication & Forgot Password
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Development server
paths:
  /accounts/authenticate:
    post:
      summary: Authenticate user credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '200':
          description: Authentication successful
        '400':
          description: Invalid credentials
  /accounts/register:
    post:
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                password:
                  type: string
                confirmPassword:
                  type: string
                acceptTerms:
                  type: boolean
              required:
                - title
                - firstName
                - lastName
                - email
                - password
                - confirmPassword
                - acceptTerms
      responses:
        '200':
          description: Registration successful
        '400':
          description: Validation error
  /accounts/verify-email:
    post:
      summary: Verify email with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
              required:
                - token
      responses:
        '200':
          description: Verification successful
        '400':
          description: Invalid token
  /accounts/forgot-password:
    post:
      summary: Send password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
              required:
                - email
      responses:
        '200':
          description: Password reset email sent
        '400':
          description: Email not found
  /accounts/reset-password:
    post:
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                password:
                  type: string
                confirmPassword:
                  type: string
              required:
                - token
                - password
                - confirmPassword
      responses:
        '200':
          description: Password reset successful
        '400':
          description: Invalid token or validation error 